Add minified build
