+++ index.js
@@ -20,6 +20,13 @@ var mixins = {
 		getStoreLink: function() {
 			return 'http://play.google.com/store/apps/details?id=' + this.appId;
 		}
+	},
+	windows: {
+		appMeta: 'msApplication-ID',
+		iconRels: ['windows-touch-icon', 'apple-touch-icon-precomposed', 'apple-touch-icon'],
+		getStoreLink: function() {
+			return 'http://www.windowsphone.com/s?appid=' + this.appId;
+		}
 	}
 };
 
@@ -32,18 +39,24 @@ var SmartBanner = function(options) {
 		button: 'OPEN', // Text for the install button
 		store: {
 			ios: 'On the App Store',
-			android: 'In Google Play'
+			android: 'In Google Play',
+			windows: 'In the Windows Store'
 		},
 		price: {
 			ios: 'FREE',
-			android: 'FREE'
+			android: 'FREE',
+			windows: 'FREE'
 		},
 		force: false // put platform type (ios, android, etc.) here for emulation
 	}, options || {});
 
 	if (this.options.force) {
 		this.type = this.options.force;
-	} else if (userAgent.match(/iPad|iPhone|iPod/i) !== null) {
+	}
+	else if (userAgent.match(/Windows Phone/i) != null && userAgent.match(/Touch/i) !== null) {
+		this.type = 'windows';
+ 	}
+ 	else if (userAgent.match(/iPad|iPhone|iPod/i) !== null) {		 	} else if (userAgent.match(/iPad|iPhone|iPod/i) !== null) {
 		if (userAgent.match(/Safari/i) !== null &&
 				(userAgent.match(/CriOS/i) !== null ||
 				Number(userAgent.substr(userAgent.indexOf('OS ') + 3, 3).replace('_', '.')) < 6)) {
@@ -141,7 +154,12 @@ SmartBanner.prototype = {
 		if (!meta) {
 			return;
 		}
-		this.appId = /app-id=([^\s,]+)/.exec(meta.getAttribute('content'))[1];
+
+		if (this.type === 'windows') {
+			this.appId = meta.getAttribute('content');
+		} else {
+			this.appId = /app-id=([^\s,]+)/.exec(meta.getAttribute('content'))[1];
+		}
 
 		return this.appId;
 	}
