+++ package.json
@@ -4,7 +4,7 @@
   "description": "Smart banner for Android or iOS",
   "main": "index.js",
   "scripts": {
-    "build": "browserify index.js -s SmartBanner > smart-app-banner.js"
+    "build": "browserify index.js -s SmartBanner | ccjs - > smart-app-banner.js"
   },
   "repository": {
     "type": "git",
@@ -40,5 +40,8 @@
     "component-query": "^0.0.3",
     "get-doc": "^1.0.3",
     "cookie-cutter": "^0.1.1"
+  },
+  "devDependencies": {
+    "closurecompiler": "^1.5.1"
   }
 }
